Add extra query to improve results for certain movie searches
Movies that are one word, such as 'Pi'(1998), 'Drive'(2011) or
'Kin'(2018) tended not to return a review using the initial query
parameters. This is because NYT api finds too many results for
such  searches (including results that may not even the words in the
their titles!). As each query only returns JSON object with
max of 20 results, the review for the searched movie often doesn't
make the first cut. To make matters more complicated, the way NYT api
returns results was sometimes confusing: queries that should have shown
the movie review did not (e.g. publication-date sorting doesn't seem to
work properly...).

After tinkering with the NYT api, I was able to discover a
reasonably sensical query that was able to find reviews for mono-word
titles. Although a second query slows down page-loading, most movie
reviews need only one query, and if not, waiting a few more seconds
for an accurate review is an acceptable trade-off.