Restrict review api to authenticated users
While JSON web tokens seem to be the standard way to do authentication
for api requests, I decided to stick with the flask_login
authentication mechanism, which I think uses cookies. Using JWT
would've meant tinkering with the original login route to pass
an access token as an argument via render_template. This seemed
ugly and messy to me and not worth the trouble. The tradeoff
here is that it makes testing from other clients such as requests, curl,
Postman, etc. more complicated, but I can live with that so long as
my test script works.